Added button "1" and button "+" functionality
The user is able to press 1 as many times as they like while being displayed in the results label.

The user is able to press + so that the procedding 1's can in turn be added.

The user is able to press + once again to retrieve the subtotal/total.

e.g
- 111 (output = 111)
- + ( output = 111)
- 11 (output = 11)
- + (output = 122)