Added the ability to +,-,*,/ numbers
Added varaible, "lastNumber", to store the last number in the form of (lastnumber operation currentnumber)

Added a varaible, "executingCal", to check if +,-,* or / was pressed.

Added a varaible, "operationTag", to store the the tag number of the operation button pressed, "+,-*,/"

New logic:

If button pressed != "" or AC or =
Store the operation sign pressed
Store the next set of numbers (executingCal=true)
If = is pressed
Based on the tag number of the operation just pressed, add the lastNumber to the currentNumber (for addition).