Added functionality of +/- button and error handling for two consecutive operations
Added the ability to make a number in display a negative number if a postive and a postitive number if a negative.

Added error handling if the user press the +,-,*,\ button twice.

Added error handling if the user press the +/- button when a +,-,*,\ button was pressed. User can only press +/- if a number is displayed.